Python Programs

+++ Odd_Even.py
@@ -1,6 +1,6 @@
 print("Program for finding the given value is odd or even")
 number = input("Enter a value : ")
-if(float(number) %2 ==0):
+if(int(number) % 2 ==0):
     print("The given number is an even number")
 else:
     print("The given number is an odd number")
